Move properties to package scope
