+++ Tests/PlayerTests/PulseTests.swift
@@ -1,0 +1,15 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+@testable import Player
+
+import CoreMedia
+import Nimble
+import XCTest
+
+final class PulseTests: XCTestCase {
+
+}
